+++ src/components/table/tableContext.ts
@@ -1,6 +1,6 @@
 import { createContext, useContext } from 'react';
 
-export const TableContext = createContext({ hasBorder: true, color: '' });
+export const TableContext = createContext({ color: '' });
 export function useTableContext() {
   const context = useContext(TableContext);
   return context;