+++ tests/app/toolbar.test.ts
@@ -1,6 +1,6 @@
-import { test, expect } from '@playwright/test';
+import { expect, test } from '@playwright/test';
 
 test('toolbar', async ({ page }) => {
   await page.goto('http://localhost:5173/pages/demo.html');
-  await expect(page.locator('_react=ToolbarItem')).toHaveCount(5);
+  await expect(page.locator('_react=ToolbarItem')).toHaveCount(6);
 });