+++ CHANGELOG.md
@@ -1,5 +1,31 @@
 # Changelog
 
+## [4.0.0](https://github.com/zakodium-oss/react-science/compare/v3.1.0...v4.0.0) (2024-05-17)
+
+
+### ⚠ BREAKING CHANGES
+
+* remove hasBorder prop from Table component
+
+### Features
+
+* add logger components ([#709](https://github.com/zakodium-oss/react-science/issues/709)) ([30b6212](https://github.com/zakodium-oss/react-science/commit/30b6212d3c87f5f5c88eae5685e8eaf110785f8b))
+* provide only one way to add borders to Table ([90cc00e](https://github.com/zakodium-oss/react-science/commit/90cc00e59a416c1222f01a90361d9a38ffe3ea2f))
+* tooltip help content ([#715](https://github.com/zakodium-oss/react-science/issues/715)) ([287f1c3](https://github.com/zakodium-oss/react-science/commit/287f1c3eff668e687ca5a19b04d23fd42185b6bf))
+
+
+### Bug Fixes
+
+* update selected dependencies ([#724](https://github.com/zakodium-oss/react-science/issues/724)) ([7452982](https://github.com/zakodium-oss/react-science/commit/7452982f3ccec44e8c4ba6f95465ed50be093711))
+
+## [3.1.0](https://github.com/zakodium-oss/react-science/compare/v3.0.0...v3.1.0) (2024-04-09)
+
+
+### Features
+
+* spread Table.Row props into HTML row tag ([#718](https://github.com/zakodium-oss/react-science/issues/718)) ([e02f013](https://github.com/zakodium-oss/react-science/commit/e02f01370e1adb43ec63feaa1ae51b18449749bb))
+* toolbar in accordion ([#698](https://github.com/zakodium-oss/react-science/issues/698)) ([93f8f45](https://github.com/zakodium-oss/react-science/commit/93f8f45afc197a013c9038ebf4829067c4e54e11))
+
 ## [3.0.0](https://github.com/zakodium-oss/react-science/compare/v2.0.0...v3.0.0) (2024-04-03)